Fixed chapter 1 errors (remaining)
Prop 1.5: repeat of h: X->Z is a bijection, fixed
Def 1.25: refer of Def 2.22 and 2.23 fixed to 1.22 and 1.23
Prop 1.9: Supposed to be proof of Prop 1.8. Fixed. Also, ref of Prop 2.7 fixed to 1.7
Prop 1.11: finite seq of countably sets -> finite seq of countable sets
Prop 1.11 Proof: ref of Prop 2.9 fixed to 1.8, Pop 2.10 fixed to prop 1.9